Completed contour representation in Path. Added tests.
