+++ CMakeLists.txt
@@ -40,6 +40,8 @@ set(SOURCES ${SOURCES} Source/Geometry/HCSize.c)
 set(SOURCES ${SOURCES} Source/Geometry/HCRectangle.c)
 set(SOURCES ${SOURCES} Source/Graphic/HCColor.c)
 set(SOURCES ${SOURCES} Source/Graphic/HCRaster.c)
+set(SOURCES ${SOURCES} Source/Graphic/HCRaster+Draw.c)
+set(SOURCES ${SOURCES} Source/Graphic/HCRaster+File.c)
 set(SOURCES ${SOURCES} Source/Graphic/HCContour.c)
 set(SOURCES ${SOURCES} Source/Graphic/HCPath.c)
 set(SOURCES ${SOURCES} Source/Graphic/HCPath+Shapes.c)
