+++ Source/Graphic/HCColor.h
@@ -86,4 +86,11 @@ HCColor HCColorScale(HCColor color, HCReal scalar);
 HCColor HCColorClamp(HCColor color);
 HCColor HCColorInterpolate(HCColor color, HCColor other, HCReal t);
 
+//----------------------------------------------------------------------------------------------------------------------------------
+// MARK: - Combinations
+//----------------------------------------------------------------------------------------------------------------------------------
+HCColor HCColorCombine(HCColor c0, HCColor c1, HCReal t);
+HCColor HCColorCombine3(HCColor ca, HCReal ta, HCColor cb, HCReal tb, HCColor cc, HCReal tc);
+HCColor HCColorCombine4(HCColor ca, HCReal ta, HCColor cb, HCReal tb, HCColor cc, HCReal tc, HCColor cd, HCReal td);
+
 #endif /* HCColor_h */
