+++ Source/Graphic/HCRaster+Draw.h
@@ -1,0 +1,45 @@
+//
+//  HCRaster+Draw.h
+//  HollowCore
+//
+//  Created by Matt Stoker on 4/18/20.
+//  Copyright © 2020 HollowCore. All rights reserved.
+//
+
+#ifndef HCRaster_Draw_h
+#define HCRaster_Draw_h
+
+#include "HCRaster.h"
+
+//----------------------------------------------------------------------------------------------------------------------------------
+// MARK: - Definitions
+//----------------------------------------------------------------------------------------------------------------------------------
+#define HCRasterColorRotatingStatic {.a = -1000000.0, .r = 0.0, .g = 0.0, .b = 0.0}
+static const HCColor HCRasterColorRotating = HCRasterColorRotatingStatic;
+
+//----------------------------------------------------------------------------------------------------------------------------------
+// MARK: - Line Drawing Operations
+//----------------------------------------------------------------------------------------------------------------------------------
+void HCRasterDrawPoint(HCRasterRef self, HCReal x, HCReal y, HCColor color);
+void HCRasterDrawLine(HCRasterRef self, HCReal x0, HCReal y0, HCReal x1, HCReal y1, HCColor c0, HCColor c1);
+void HCRasterDrawQuadraticCurve(HCRasterRef self, HCReal x0, HCReal y0, HCReal cx, HCReal cy, HCReal x1, HCReal y1, HCColor c0, HCColor c1);
+void HCRasterDrawCubicCurve(HCRasterRef self, HCReal x0, HCReal y0, HCReal cx0, HCReal cy0, HCReal cx1, HCReal cy1, HCReal x1, HCReal y1, HCColor c0, HCColor c1);
+void HCRasterDrawArc(HCRasterRef self, HCReal x0, HCReal y0, HCReal x1, HCReal y1, HCReal xr, HCReal yr, HCReal theta, HCBoolean largeArc, HCBoolean sweep, HCColor c0, HCColor c1);
+
+//----------------------------------------------------------------------------------------------------------------------------------
+// MARK: - Shape Drawing Operations
+//----------------------------------------------------------------------------------------------------------------------------------
+void HCRasterDrawTriangle(HCRasterRef self, HCReal ax, HCReal ay, HCReal bx, HCReal by, HCReal cx, HCReal cy, HCColor ca, HCColor cb, HCColor cc);
+void HCRasterFillTriangle(HCRasterRef self, HCReal ax, HCReal ay, HCReal bx, HCReal by, HCReal cx, HCReal cy, HCColor ca, HCColor cb, HCColor cc);
+void HCRasterFillTexturedTriangle(HCRasterRef self, HCReal ax, HCReal ay, HCReal bx, HCReal by, HCReal cx, HCReal cy, HCRasterRef texture, HCReal tax, HCReal tay, HCReal tbx, HCReal tby, HCReal tcx, HCReal tcy);
+void HCRasterDrawQuad(HCRasterRef self, HCReal ax, HCReal ay, HCReal bx, HCReal by, HCReal cx, HCReal cy, HCReal dx, HCReal dy, HCColor ca, HCColor cb, HCColor cc, HCColor cd);
+void HCRasterFillQuad(HCRasterRef self, HCReal ax, HCReal ay, HCReal bx, HCReal by, HCReal cx, HCReal cy, HCReal dx, HCReal dy, HCColor ca, HCColor cb, HCColor cc, HCColor cd);
+void HCRasterFillTexturedQuad(HCRasterRef self, HCReal ax, HCReal ay, HCReal bx, HCReal by, HCReal cx, HCReal cy, HCReal dx, HCReal dy, HCRasterRef texture, HCReal tax, HCReal tay, HCReal tbx, HCReal tby, HCReal tcx, HCReal tcy, HCReal tdx, HCReal tdy);
+
+//----------------------------------------------------------------------------------------------------------------------------------
+// MARK: - Path Drawing Operations
+//----------------------------------------------------------------------------------------------------------------------------------
+void HCRasterDrawPath(HCRasterRef self, HCPathRef path, HCColor color);
+void HCRasterDrawPathData(HCRasterRef self, const char* pathData, HCColor color);
+
+#endif /* HCRaster_Draw_h */
